+++ request.js
@@ -10,6 +10,7 @@ function request(url, opts) {
 
   var req = Web.request(assign({}, opts, url))
   if (opts && "body" in opts) write(req, opts.body)
+  if (opts && "timeout" in opts && opts.timeout > 0) timeout(req, opts.timeout)
   req.end()
 
   return new Promise(resolve.bind(null, req))
@@ -17,6 +18,7 @@ function request(url, opts) {
 
 function resolve(req, resolve, reject) {
   req.once("response", resolve)
+  req.once("timeout", reject)
   req.once("error", reject)
 }
 
@@ -29,6 +31,23 @@ function write(req, body) {
     case "undefined": break
     default: throw new TypeError("Invalid body type: " + type)
   }
+}
+
+function timeout(req, timeout) {
+  // Start counting timeout from the moment a connection can be made.
+  // When using an overloaded HTTP Agent, an actual connection might be delayed.
+  req.on("socket", function() {
+    var id = setTimeout(timeoutAbort, timeout, req)
+    req.on("error", clearTimeout.bind(null, id))
+    req.on("response", clearTimeout.bind(null, id))
+  })
+}
+
+function timeoutAbort(req) {
+  req.abort(req)
+  var err = new Error("Fetch Timeout")
+  err.code = "ETIMEDOUT"
+  req.emit("timeout", err)
 }
 
 function typeOf(value) { return value === null ? "null" : typeof value }